requirements.txt
openai
semantic-kernel #1.23.1
ipykernel
python-dotenv


